Tidy code and files
